requirements.txt
streamlit
PyMuPDF
python-docx
spacy
scikit-learn
requests
pandas
Pillow
